+++ CHANGES.md
@@ -2,7 +2,9 @@
 
 ## 2.34.0
 
-This release add support for `--pip-version 25.1`.
+This release add support for `--pip-version 25.1` as well as `--pip-version latest-compatible`. The
+`latest-compatible` version will be the latest `--pip-version` supported by Pex compatible with the
+current interpreter running Pex.
 
 * Add support for `--pip-version 25.1`. (#2744)